src/requirements.txt
torch
torchvision
torch-summary
lightning
transformers
datasets
sentencepiece
Pillow
argparse
numpy
wandb

